+++ .gitignore
@@ -15,9 +15,3 @@
 #composer req twig
 #composer req asset
 #composer req doctrine
-###> symfony/webpack-encore-bundle ###
-/node_modules/
-/public/build/
-npm-debug.log
-yarn-error.log
-###< symfony/webpack-encore-bundle ###
