+++ composer.json
@@ -10,24 +10,17 @@
         "doctrine/doctrine-bundle": "^2.8",
         "doctrine/doctrine-migrations-bundle": "^3.2",
         "doctrine/orm": "^2.14",
-        "sensio/framework-extra-bundle": "^6.2",
         "symfony/console": "6.2.*",
         "symfony/dotenv": "6.2.*",
-        "symfony/flex": "^2.2",
-        "symfony/form": "6.2.*",
+        "symfony/flex": "^2",
         "symfony/framework-bundle": "6.2.*",
         "symfony/maker-bundle": "^1.48",
         "symfony/runtime": "^6.2",
         "symfony/security-bundle": "6.2.*",
-        "symfony/security-csrf": "6.2.*",
         "symfony/twig-bundle": "6.2.*",
-        "symfony/validator": "6.2.*",
-        "symfony/webpack-encore-bundle": "^1.16",
         "symfony/yaml": "6.2.*",
         "twig/extra-bundle": "^2.12|^3.0",
-        "twig/twig": "^2.12|^3.0",
-        "ext-http": "*",
-        "ext-json/ext-json": "*"
+        "twig/twig": "^2.12|^3.0"
     },
     "config": {
         "allow-plugins": {
@@ -76,11 +69,5 @@
             "allow-contrib": false,
             "require": "6.2.*"
         }
-    },
-    "require-dev": {
-        "symfony/debug-bundle": "6.2.*",
-        "symfony/monolog-bundle": "^3.0",
-        "symfony/stopwatch": "6.2.*",
-        "symfony/web-profiler-bundle": "6.2.*"
     }
 }
