--- src/Form/RegistrationFormType.php
@@ -1,55 +1,0 @@
-<?php
-
-namespace App\Form;
-
-use App\Entity\Credentials;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
-use Symfony\Component\Form\Extension\Core\Type\PasswordType;
-use Symfony\Component\Form\Extension\Core\Type\RepeatedType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\Component\Validator\Constraints\IsTrue;
-use Symfony\Component\Validator\Constraints\Length;
-use Symfony\Component\Validator\Constraints\NotBlank;
-
-class RegistrationFormType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('Pseudo')
-            ->add('plainPassword', RepeatedType::class, [
-                // instead of being set onto the object directly,
-                // this is read and encoded in the controller
-                'type' => PasswordType::class,
-                'invalid_message' => 'Le mot de passe doit être identique',
-                'mapped' => false,
-                'attr' => ['autocomplete' => 'new-password'],
-                'options' => ['attr' => ['class' => 'password-field']],
-                'required' => true,
-                'first_options'  => ['label' => 'Mot de passe'],
-                'second_options' => ['label' => 'Répètez le Mot de passe'],
-                'constraints' => [
-                    new NotBlank([
-                        'message' => 'Merci de saisir un mot de passe',
-                    ]),
-                    new Length([
-                        'min' => 4,
-                        'minMessage' => 'Un mot de passe ne peux avoir plus de {{ limit }} caractères',
-                        // max length allowed by Symfony for security reasons
-                        'max' => 255,
-                        'maxMessage' => 'Un mot de passe ne peux dépasser {{ limit }} caractères'
-                    ]),
-                ],
-            ])
-        ;
-    }
-
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefaults([
-            'data_class' => Credentials::class,
-        ]);
-    }
-}
