+++ templates/event/index.html.twig
@@ -3,5 +3,5 @@
 {% block title %}Hello EventController!{% endblock %}
 
 {% block body %}
-<h1> coucou </h1>
+
 {% endblock %}
