--- templates/user/index.html.twig
@@ -1,57 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}User index{% endblock %}
-
-{% block body %}
-
-    <div>
-  <a href="{{ path('app_user_index') }}">Home</a>
-  <a href="{{ path('app_user_new') }}">Create a new user</a><br>
-    </div>
-
-
-    <div>
-    <label for="SearchUser">Find a User:</label>
-    <input type="search" id= "SearchUser" name="SearchUser" placeholder="Pseudo..." >
-    <button>Search</button>
-    </div>
-
-     <h1>User index</h1>
-
-    <table class="table">
-        <thead>
-            <tr>
-                <th>Id</th><br>
-                <th>Name</th>
-                <th>Surname</th>
-                <th>Phone</th>
-                <th>Email</th>
-
-                <th>actions</th>
-            </tr>
-        </thead>
-        <tbody>
-        {% for user in users %}
-            <tr>
-                <td>{{ user.id }}</td>
-                <td>{{ user.name }}</td>
-                <td>{{ user.surname }}</td>
-                <td>{{ user.phone }}</td>
-                <td>{{ user.email }}</td>
-
-                <td>
-                    <a href="{{ path('app_user_show', {'id': user.id}) }}">show</a>
-                    <a href="{{ path('app_user_edit', {'id': user.id}) }}">edit</a>
-                    <a href="{{ path('app_user_delete', {'id': user.id}) }}">delete</a>
-                </td>
-            </tr>
-        {% else %}
-            <tr>
-                <td colspan="8">no records found</td>
-            </tr>
-        {% endfor %}
-        </tbody>
-    </table>
-
-    <a href="{{ path('app_user_new') }}">Create new</a>}}}
-{% endblock %}
