--- templates/user/new.html.twig
@@ -1,30 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}New User{% endblock %}
-
-{% block body %}
-
-    <h1>Create new User</h1>
-
-{#  {{ include('registration/register.html.twig') }}#}
-
-    {{ form_start(form) }}
-    {{ form_widget(form) }}
-    <button>Valider</button>
-    {{ form_end(form) }}
-
-
-    <a href="{{ path('app_user_index') }}">back to list</a>
-
-
-    {#
-    {{ form_start(registrationForm) }}
-    {{ form_row(registrationForm.pseudo) }}
-    {{ form_row(registrationForm.plainPassword, {
-        label: 'Password'
-    }) }}
-    {{ form_row(registrationForm.agreeTerms) }}
-    #}
-
-
-{% endblock %}
